requirements.txt
alembic==1.7.5
attrs==21.2.0
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
coverage==6.1.2
cycler==0.11.0
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
fonttools==4.28.5
greenlet==2.0.1
gevent==22.10.2
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
Mako==1.2.2
MarkupSafe==2.0.1
matplotlib==3.5.1
mpld3==0.5.7
multidict==5.2.0
numpy==1.22.0
packaging==21.3
pandas==1.3.5
Pillow==9.0.1
pluggy==1.0.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
py==1.11.0
PyJWT==2.4.0
pyparsing==3.0.6
pytest==6.2.5
pytest-cov==3.0.0
pytest-vcr==1.0.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
scipy==1.7.3
six==1.16.0
SQLAlchemy==1.4.27
toml==0.10.2
tomli==1.2.2
urllib3==1.26.7
vcrpy==4.1.1
Werkzeug==2.0.2
wrapt==1.13.3
WTForms==3.0.0
yarl==1.7.2
